requirements.txt
discord.py
asyncpraw
python-telegram-bot
Flask
pyping
mcstatus
pysftp
mcrcon